+++ modules/ZWaveGate/classes/ZWaveBatteryDevice.js
@@ -17,7 +17,9 @@ ZWaveBatteryDevice = function (id, controller, zDeviceId, zInstanceId) {
 
     this.deviceType = "probe";
     this.deviceSubType = "battery";
+    this.resetIconBase();
 
+    this.resetIconBase();
     this.widgetClass = "ProbeWidget";
 
     this.setMetricValue("probeTitle", "Battery");