+++ modules/ZWaveGate/classes/ZWaveDoorlockDevice.js
@@ -16,6 +16,7 @@ ZWaveDoorlockDevice = function (id, controller, zDeviceId, zInstanceId) {
     this.zCommandClassId = 0x62;
 
     this.deviceType = "doorlock";
+    this.resetIconBase();
 
     this.widgetClass = "DoorlockWidget";